Modify FirstMenuActivity UI & Auto Login Function
* Modify FirstMenuActivity Button UI
* Auto Login Funtion move from MainActivity to SplashActivity
