+++ app/src/main/java/io/kong/incheon/nfc_check/activity/FirstMenuActivity.java
@@ -6,6 +6,7 @@ import android.os.Bundle;
 import android.support.v7.app.AppCompatActivity;
 import android.view.View;
 import android.widget.Button;
+import android.widget.ImageButton;
 
 import io.kong.incheon.nfc_check.R;
